pyproject.toml
[build-system]
requires = ["hatchling>=1.27"]
build-backend = "hatchling.build"

[project]
name = "foga"
version = "0.1.0"
description = "Unified developer CLI for Python packages with C and C++ bindings."
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Federico Lucco Castello" }
]
dependencies = [
    "PyYAML>=6.0",
    "typer>=0.16",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0",
]
dev = [
    "build>=1.2",
    "pytest>=8.0",
    "ruff>=0.11",
    "tox>=4.0",
    "twine>=5.0",
]

[project.scripts]
foga = "foga.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/foga"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py310"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I"]
